# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Tradosphere OS Gateway API
  version: "1.0.0"
  description: |-
    Unified API gateway for Tradosphere OS (Sprint 9, task 9.4). Implements
    the architecture decided in full in Decision D19 (EXECUTION_BOOK.md)
    before any gateway code was written.

    Two route families exist side by side:

    - Proxied (Auth, Market Data, Education, Portfolio, Analytics): the
      gateway reverse-proxies to each service's existing HTTP API via a
      hand-rolled wildcard handler that forwards the Authorization header
      verbatim. Each upstream service remains the sole auth authority for
      its own routes -- the gateway does NOT re-authenticate proxied
      requests.
    - In-process (Research, AI Council, CIO, Paper Trading, Journal): these
      five services are library-only with no HTTP surface of their own, so
      their routes are implemented directly on the gateway and authenticated
      by the gateway's own requireAuth (identical factory pattern to every
      other service in this repo).

    All domain routes are versioned under /v1/<service>/.... Infra routes
    (/health, /health/services, /metrics, /stream, /documentation,
    /openapi.yaml) are unversioned and, along with auth's own signup/login,
    are the only unauthenticated routes in the gateway.
servers:
  - url: http://localhost:4000
    description: Local gateway (apps/api)
...
security:
  - bearerAuth: []
paths:
...
  /v1/analytics/win-rate:
    get:
      tags: [Analytics]
      operationId: analyticsGetWinRate
      summary: Win rate over an optional date range
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Win rate computed fresh from the caller's own journal entries.
          content:
            application/json:
              schema:
                type: object
                properties:
                  winRate:
                    type: number
                required:
                  - winRate
        '400':
          description: Invalid from/to timestamp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/analytics/average-return:
    get:
      tags: [Analytics]
      operationId: analyticsGetAverageReturn
      summary: Average return (absolute and percentage) over an optional date range
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Average return computed fresh from the caller's own journal entries.
          content:
            application/json:
              schema:
                type: object
                properties:
                  averageReturn:
                    type: number
                  averageReturnPct:
                    type: number
                required:
                  - averageReturn
                  - averageReturnPct
        '400':
          description: Invalid from/to timestamp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/analytics/risk-reward:
    get:
      tags: [Analytics]
      operationId: analyticsGetRiskReward
      summary: Planned and realized risk:reward ratios over an optional date range
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Risk:reward ratios computed fresh from the caller's own journal entries.
          content:
            application/json:
              schema:
                type: object
                properties:
                  plannedRiskRewardRatio:
                    type: number
                  realizedRiskRewardRatio:
                    type: number
                required:
                  - plannedRiskRewardRatio
                  - realizedRiskRewardRatio
        '400':
          description: Invalid from/to timestamp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/analytics/expectancy:
    get:
      tags: [Analytics]
      operationId: analyticsGetExpectancy
      summary: Trade expectancy over an optional date range
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Expectancy computed fresh from the caller's own journal entries.
          content:
            application/json:
              schema:
                type: object
                properties:
                  expectancy:
                    type: number
                required:
                  - expectancy
        '400':
          description: Invalid from/to timestamp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/analytics/drawdown:
    get:
      tags: [Analytics]
      operationId: analyticsGetDrawdown
      summary: Maximum drawdown percentage over an optional date range
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Max drawdown computed fresh from the caller's own portfolio snapshots.
          content:
            application/json:
              schema:
                type: object
                properties:
                  maxDrawdownPct:
                    type: number
                required:
                  - maxDrawdownPct
        '400':
          description: Invalid from/to timestamp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/analytics/risk-adjusted-returns:
    get:
      tags: [Analytics]
      operationId: analyticsGetRiskAdjustedReturns
      summary: Sharpe and Sortino ratios over an optional date range
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Risk-adjusted returns computed fresh from the caller's own portfolio snapshots. Risk-free rate defaults to 0 (no rate feed exists).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskAdjustedReturns'
        '400':
          description: Invalid from/to timestamp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/analytics/performance:
    get:
      tags: [Analytics]
      operationId: analyticsGetPerformance
      summary: Combined rollup of every granular stat over an optional date range
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: The same computeFullStatSet() result POST /v1/analytics/reports would persist at this moment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullStatSet'
        '400':
          description: Invalid from/to timestamp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/analytics/monthly-reports:
    get:
      tags: [Analytics]
      operationId: analyticsGetMonthlyReports
      summary: Stats broken down by UTC calendar month
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: One entry per UTC calendar month with at least one trade filled in it.
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items:
                      $ref: '#/components/schemas/MonthlyReport'
                required:
                  - reports
        '400':
          description: Invalid from/to timestamp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/analytics/strategy-stats:
    get:
      tags: [Analytics]
      operationId: analyticsGetStrategyStats
      summary: Stats grouped by strategy proxy (CIO verdict label + recommended direction)
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Largest group first; ties broken alphabetically by strategy key.
          content:
            application/json:
              schema:
                type: object
                properties:
                  strategies:
                    type: array
                    items:
                      $ref: '#/components/schemas/StrategyStats'
                required:
                  - strategies
        '400':
          description: Invalid from/to timestamp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/analytics/trade-distribution:
    get:
      tags: [Analytics]
      operationId: analyticsGetTradeDistribution
      summary: Closed-trade realized P&L distribution across equal-width buckets
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: buckets
          in: query
          required: false
          description: Overrides the default of 10 equal-width buckets.
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Bucket boundaries are derived from the caller's own real min/max realized P&L, never a fixed hardcoded threshold.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradeDistribution'
        '400':
          description: Invalid from/to timestamp or buckets value.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/analytics/heatmap:
    get:
      tags: [Analytics]
      operationId: analyticsGetHeatmap
      summary: Day-of-week x session-window performance heatmap
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Always 28 cells (7 days x 4 sessions), even when a cell has zero trades.
          content:
            application/json:
              schema:
                type: object
                properties:
                  cells:
                    type: array
                    items:
                      $ref: '#/components/schemas/HeatmapCell'
                required:
                  - cells
        '400':
          description: Invalid from/to timestamp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationFailure'
        '401':
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |-
        Every proxied service verifies this itself against its own JWT
        secret; the gateway's own requireAuth (used only by the 10
        in-process routes) uses the identical verifyAccessToken /
        InvalidTokenError contract from @tradosphere/auth as every other
        service.
  schemas:
    ErrorResponse:
      type: object
      description: |-
        The generic shape every service's setErrorHandler falls back to for
        unexpected errors, and the shape most caught domain errors use
        (sometimes with extra fields alongside error, e.g. 409 responses
        that also carry missingPriceSymbols). Not a Fastify-default
        {statusCode,error,message} envelope -- just {error, ...}.
      properties:
        error:
          type: string
      required: [error]
    ValidationFailure:
      type: object
      description: The zod-failure shape validateBody() returns on every 400, identical across every service (reused verbatim from services/education's original).
      properties:
        error:
          type: string
          example: Validation failed
        details:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
            required: [path, message]
      required: [error, details]
...